requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Pillow==10.1.0
python-dotenv==1.0.0
qrcode==7.4.2
werkzeug==2.3.7
gunicorn==21.2.0 
authlib==1.2.1
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 